PC ASSEMBLE (LAB)
On 12 Dec 2023, our class commit a PC ASSEMBLE LAB session. Each of the groups were given a dissembled computer case, and then the instructors taught us the way to assemble the case hand by hand. Me and my team eventually successful to assemble the computer case. The instructors and the crews was hearty to give support to us. This is my first time assembling the computer case. I really appreciate the opportunity to have a different way to learn about computer. I think this lab session will have a gigantic positive impact to my future!